requirements.txt
coverage==7.2.5
Django==4.2.5
django-nose==1.4.7
djangorestframework==3.15.2
nose==1.3.7
psycopg2==2.9.7
